requirements.txt
Appium-Python-Client>=5.0.0
pytest>=8.0.0
pytest-html>=4.0.0
allure-pytest>=2.13.0
selenium>=4.20.0
